It has been done, the normal mapping bug has been fixed
I always assumed that my texture abstraction was working correctly and was initializing the image views with the VkFormat provided to the constructor. This was not the case. All texture view were created with VK_FORMAT_R8G8B8_SRGB, while normal maps need to be specified as UNORM. Fixing this, fixed the issue.